LG-1112 Use a base class for recuring job controllers (#2864)
**Why**: To DRY up the code a bit and to make sure that we are using secure compare everywhere we a checking API auth tokens.
